Fix copy construction and assignment of Vm
If you just copy it by the default copy constructor, the pointers in Cpu
to the Stack and Source will be wrong. So implement copy manually.

If anyone has questions about the rule of five:

I didn't bother to implement move, since we can just fall back on
copies. Performance is not important yet.

The default constructor actually works just fine, so don't bother with
that.